Fixed minor things that are hopefully not controversial
It reads well, which is nice considering the length.  A few styistic points:

A number of paragraphs consist of a single sentence. While that's not unforgivable, there's a decent chance they'd feel a bit snappier split up.

I have a personal dislike for adjacent sentences featuring repeated adverbs e.g. "quickly sets about... quickly start tearing themselves apart".  Sometimes this feels like it might be intentional, but other times it seems excessive. Maybe just cut some adverbs?

There is also some divergence in tense and tone between different parts. e.g. "And so it was that we..." with its psuedo-religious text feel is a bit of a surprise immediately after "This realism combined with his bleak outlook on life makes him a somewhat unique character in the roster of hollywood heroes."

I think you're particularly strong on the part about Pitch Black, and unfortunately just as you get off it, the reader is starting to tire from the length, just as the passion in your writing is flagging a little too. I think for online presentation, you could consider publishing them as separate pages / chapters with links at the bottom to the next.  I enjoy the 'alternate history' ideas, and wonder if it might be good to set them out from the text using color or formatting, maybe even having a few more of them if they're short and the main body text can flow around them.

At the beginning you set yourself the task of explaining *why* the Riddick franchise doesn't understand itself, but you don't actually do much of that.  You actually spend your words 1. Explaining the true meaning of the first film, and 2. showing that neither the second nor the third live up to it.

Basically, I think your analysis of Pitch Black is superb, could even perhaps be explored in more detail. While it does explain why the next two films are so disappointing, that in itself isn't interesting enough to carry too many paragraphs.  How about this for a structure: 1. Why hollywood thinks pitch black was great (with quotes if you can find them). 2. Why Pitch Black was really great. 3. Is it possible to make good sequels to a film like pitch black?  What could they look like?